fixing crud for client

+++ backend/app/routes/users.py
@@ -1,6 +1,7 @@
 from typing import List
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException, Response
 from sqlalchemy import select
+from starlette.status import HTTP_204_NO_CONTENT, HTTP_404_NOT_FOUND
 from app import models, schemas
 from app.auth import get_current_user
 from app.database import get_db
@@ -12,38 +13,41 @@ router = APIRouter(prefix="/users", tags=["Users"])
 
 
 @router.post("/")
-async def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
-    new_client = models.Client(**user.model_dump())
+async def create_user(client: schemas.UserCreate, db: Session = Depends(get_db)):
+    new_client = models.Client(**client.model_dump())
     new_client.typ = "client"
-    new_client.subscription_model = "free"
     db.add(new_client)
     await db.commit()
     await db.refresh(new_client)
-    return {"user": new_client.email}
+    return {"client": new_client.email}
 
 
 @router.get("/", response_model=List[schemas.UserOut])
 async def get_users(db: Session = Depends(get_db)):
-    users = await db.execute(select(models.User))
+    users = await db.execute(select(models.Client))
     return users.scalars().all()
 
 
 @router.patch("/", response_model=schemas.UserOut)
 async def update_user(
     updated_user: schemas.UserUpdate,
-    current_user: int = Depends(get_current_user),
+    current_client: int = Depends(get_current_user),
     db: Session = Depends(get_db),
 ):
-    user = current_user
+    if not client:
+        raise HTTPException(HTTP_404_NOT_FOUND, detail="user does not exist")
+
+    client = current_client
 
     user_dict = updated_user.model_dump(exclude_none=True)
 
     for key, value in user_dict.items():
-        setattr(user, key, value)
+        setattr(client, key, value)
 
     await db.commit()
-    await db.refresh(user)
-    return user
+    await db.refresh(client)
+    return client
+
 
 
 @router.delete("/")
@@ -51,6 +55,9 @@ async def delete_user(
     current_user: models.User = Depends(get_current_user),
     db: Session = Depends(get_db),
 ):
+    if not current_user:
+        raise HTTPException(HTTP_404_NOT_FOUND, detail="user doesn't exist")
+
     await db.delete(current_user)
     await db.commit()
     return Response(status_code=HTTP_204_NO_CONTENT)
@@ -58,5 +65,6 @@ async def delete_user(
 
 @router.get("/profile", response_model=schemas.UserOut)
 async def get_profile(current_user: int = Depends(get_current_user)):
-    print(current_user.phone_number)
+    if not current_user:
+        raise HTTPException(HTTP_404_NOT_FOUND, detail="user doesn't exist")
     return current_user
